+++ app/model/Order.ts
@@ -18,7 +18,7 @@ export const Order = sequelize.define('Order', {
         allowNull: false
     },
     quantity: {
-        type: DataTypes.INTEGER(),
+        type: DataTypes.DECIMAL(10,2),
         allowNull: false
     },
     status: {
